+++ R/AAA_parameters.R
@@ -33,39 +33,28 @@
   if (is.null(defaults$verbose))
     defaults$verbose <- FALSE
 
+  o <- integer()
   if (length(parameter) != 0) {
     o <- pmatch(names(parameter), names(defaults))
 
     ## unknown parameter
     if (any(is.na(o))) {
       warning(sprintf(
-        ngettext(
-          length(is.na(o)),
-          "Unknown parameter: %s",
-          "Unknown parameters: %s"
-        ),
+        "Unknown control parameter(s): %s",
         paste(names(parameter)[is.na(o)],
-          collapse = ", ")
+              collapse = ", ")
       ),
         call. = FALSE,
         immediate. = TRUE)
-
-      cat("Available parameter (with default values):\n")
-      cat(rbind(names(defaults), " = ", gsub("\n", " ", as.character(defaults))),
-        sep = c("\t", " ", "\n"))
     }
 
+    ### defaults are now the actual parameters
     defaults[o[!is.na(o)]] <- parameter[!is.na(o)]
   }
 
-  if (defaults$verbose) {
-    cat("Used parameters:\n")
-    cat(rbind(names(defaults), " = ",
-      strtrim(gsub(
-        "\n", " ", as.character(defaults)
-      ), 50)),
-      sep = c("\t", " ", "\n"))
-    cat("\n")
+  if (defaults$verbose || any(is.na(o))) {
+    cat("control:\n")
+    .print_control(defaults, "used values")
   }
 
   defaults